Added Random.java utility file Added Items.java file for storing arrays of item types Converted tile and coordinate positions to use integer instead of double as Bukkit only uses integers for teleporting Added Chest randomization to Hungergames minigame, need to change chest coordinates to load chests.

+++ src/com/aboundedskull/minigames/data/Tile.java
@@ -2,27 +2,27 @@ package com.aboundedskull.minigames.data;
 
 public class Tile {
 
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
-    public Tile(double x, double y){
+    public Tile(int x, int y){
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 }
